requirements.txt
requests>=2.31.0
pandas>=2.2.0
python-dotenv>=1.0.1